Windows/requirements.txt
numpy>=2.0.1
playwright>=1.45.1
pyzbar>=0.1.9
pillow>=10.4.0
pyvirtualdisplay>=3.0
pywhatkit>=5.4.0
qrcode>=7.4.2
selenium>=4.9.1
webdriver-manager>=4.0.2